rename addliquidity form 2
